chore(factorial/consolidate): Factorial Tests and Question All consolidated.
